requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
pyjwt
mysqlclient
python-dotenv